python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentk8"
version = "2.3.0"
description = "Multi-LLM Council Backend for AGENT-K - Three-stage consensus with GPT, Gemini, and Claude"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Aditya Katiyar", email = "aditya@example.com"}
]
keywords = [
    "claude",
    "claude-code",
    "gpt",
    "gemini",
    "ai",
    "agents",
    "multi-agent",
    "llm",
    "council",
    "consensus",
    "litellm",
    "cli",
    "terminal",
    "developer-tools",
    "ml",
    "machine-learning"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.50.0",
    "python-dotenv>=1.0.0",
    "pathspec>=0.12.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/de5truct0/agentk"
Documentation = "https://github.com/de5truct0/agentk#readme"
Repository = "https://github.com/de5truct0/agentk.git"
Issues = "https://github.com/de5truct0/agentk/issues"

[project.scripts]
agentk-council = "agentk.council:main"

[tool.setuptools]
packages = ["agentk"]
include-package-data = true
